Rename storage_sync to remote_timeline_client
