+++ pageserver/src/lib.rs
@@ -13,7 +13,6 @@ pub mod profiling;
 pub mod repository;
 pub mod task_mgr;
 pub mod tenant;
-
 pub mod trace;
 pub mod virtual_file;
 pub mod walingest;
